Admin refactoring
