+++ includes/admin/class-plugin-list-table.php
@@ -10,7 +10,7 @@
  * @license https://www.gnu.org/licenses/gpl-2.0.html
  */
 
-namespace WP_Autoplugin;
+namespace WP_Autoplugin\Admin;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -87,9 +87,10 @@ class Plugin_List_Table extends \WP_List_Table {
 			$actions['activate'] = sprintf( '<a href="%s">%s</a>', esc_url( $url ), esc_html__( 'Activate', 'wp-autoplugin' ) );
 		}
 
-		$actions['fix']    = sprintf( '<a href="%s">%s</a>', esc_url( wp_nonce_url( admin_url( 'admin.php?page=wp-autoplugin-fix&plugin=' . $item['plugin_path'] ), 'wp-autoplugin-fix-plugin', 'nonce' ) ), esc_html__( 'Fix', 'wp-autoplugin' ) );
-		$actions['extend'] = sprintf( '<a href="%s">%s</a>', esc_url( wp_nonce_url( admin_url( 'admin.php?page=wp-autoplugin-extend&plugin=' . $item['plugin_path'] ), 'wp-autoplugin-extend-plugin', 'nonce' ) ), esc_html__( 'Extend', 'wp-autoplugin' ) );
-		$actions['delete'] = sprintf( '<a href="%s">%s</a>', esc_url( wp_nonce_url( admin_url( 'admin.php?page=wp-autoplugin&action=delete&plugin=' . $item['plugin_path'] ), 'wp-autoplugin-activate-plugin', 'nonce' ) ), esc_html__( 'Delete', 'wp-autoplugin' ) );
+		$actions['fix']     = sprintf( '<a href="%s">%s</a>', esc_url( wp_nonce_url( admin_url( 'admin.php?page=wp-autoplugin-fix&plugin=' . $item['plugin_path'] ), 'wp-autoplugin-fix-plugin', 'nonce' ) ), esc_html__( 'Fix', 'wp-autoplugin' ) );
+		$actions['extend']  = sprintf( '<a href="%s">%s</a>', esc_url( wp_nonce_url( admin_url( 'admin.php?page=wp-autoplugin-extend&plugin=' . $item['plugin_path'] ), 'wp-autoplugin-extend-plugin', 'nonce' ) ), esc_html__( 'Extend', 'wp-autoplugin' ) );
+		$actions['explain'] = sprintf( '<a href="%s">%s</a>', esc_url( wp_nonce_url( admin_url( 'admin.php?page=wp-autoplugin-explain&plugin=' . $item['plugin_path'] ), 'wp-autoplugin-explain-plugin', 'nonce' ) ), esc_html__( 'Explain', 'wp-autoplugin' ) );
+		$actions['delete']  = sprintf( '<a href="%s">%s</a>', esc_url( wp_nonce_url( admin_url( 'admin.php?page=wp-autoplugin&action=delete&plugin=' . $item['plugin_path'] ), 'wp-autoplugin-activate-plugin', 'nonce' ) ), esc_html__( 'Delete', 'wp-autoplugin' ) );
 
 		// Replicate the default Plugin List Table column rendering.
 		return sprintf( '<strong>%1$s</strong> v%2$s %3$s %4$s', esc_html( $item['Name'] ), esc_html( $item['Version'] ), '<p>' . wp_kses_post( $item['Description'] ) . '</p>', $this->row_actions( $actions ) );
