+++ includes/class-plugin-explainer.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * Autoplugin Explainer class.
+ *
+ * @package WP-Autoplugin
+ * @since 1.3
+ * @version 1.3
+ * @link https://wp-autoplugin.com
+ * @license GPL-2.0+
+ * @license https://www.gnu.org/licenses/gpl-2.0.html
+ */
+
+namespace WP_Autoplugin;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Plugin Extender class.
+ */
+class Plugin_Explainer {
+
+}
